pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "x-ai-job-monitor"
version = "0.1.0"
description = "Public-safe X AI job monitor with daily GitHub Actions execution."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27,<1",
  "PyYAML>=6.0,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
]

[tool.setuptools.packages.find]
include = ["x_job_monitor*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

